// movies/src/interfaces/forms.ts
export interface IErrorUser {
    firstName?: boolean;
    lastName?: boolean;
    email?: boolean;
    pass?: boolean;
    repeatPass?: boolean;
    captcha?: boolean;
    username?: boolean;
}

export interface IColors {
    firstName?: string;
    lastName?: string;
    email?: string;
    pass?: string;
    repeatPass?: string;
    username?: string | undefined;
}